Create useStickyPosition hook
