+++ admin/src/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useMenuData } from './use-menu-data';
 export { default as usePluginConfig } from './use-plugin-config';
+export { default as useStickyPosition } from './use-sticky-position';
